moved non-main stuff in main file into lib file
