+++ aw_watcher_window/__init__.py
@@ -1,1 +1,3 @@
+from . import lib
+
 from .main import main
